move thread local gdal options to utils
